Support 1.21 villagers and trades

+++ NbtTools/Entities/Trading/Enchantment.cs
@@ -3,9 +3,9 @@
     public class Enchantment
     {
         public string Id { get; private set; }
-        public short Level {  get; private set; }
+        public int Level {  get; private set; }
 
-        public Enchantment(string id, short level)
+        public Enchantment(string id, int level)
         {
             Id = id;
             Level = level;
